Propose adding the "Are" keyword to the list of boolean string
replacements.

Currently there doesn't seem to be a verb that's great for describing a property that belongs to a set of objects.
Adding the "Are" keyword to the list of replacements seems like it would cater for this.

Use case: The fund totals are stale
The fund totals aren't stale